Add support for FTD failover state
As described here: https://community.librenms.org/t/high-availability-on-ftd-devices/18656
The failover state for Cisco FTD devices is not available by default. The simple functional solution is to copy the file asa.inc.php to ftd.inc.php. This PR includes the ftd.inc.php file in order to publish this feature for general release.